Set has_started_first_navigation_ correctly in ExtensionFrameHelper
ExtensionFrameHelper::ReadyToCommitNavigation() watches for the first
navigation in a new Chrome Apps window to block the parsing of the new
document, so that the create() function's callback is called before
the document is loaded.

However, it can currently mistakenly think that a non-first
navigation that uses a new RenderFrame and a new ExtensionFrameHelper
is actually a first navigation, because it tracks that information
with has_started_first_navigation_, which is false on new
ExtensionFrameHelpers. New RenderFrames will be created on cross-site
navigations, or every cross-document navigation if RenderDocument is
enabled.

This CL fixes this by setting has_started_first_navigation_ to true
if it sees that the ExtensionFrameHelper is created for a frame that
had navigated before (loaded a non-initial empty document).

See also: https://groups.google.com/a/google.com/g/chrome-security-architecture/c/vcWW1xOdUjw/m/BU6e5OE7AwAJ

Bug: 936696
Change-Id: Idd6b7c87f72fd6a5b250e83490ab0f093309f249
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4311504
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1201492}
NOKEYCHECK=True
GitOrigin-RevId: ef8168e2506d4b5b08dd09fc87f28a7a4c372eb5