Move frame drop UMAs from Browser to Renderer process.
This CL moves logging/UMA about frame drops from VideoCaptureController
(Browser process) to the MediaStreamVideoTrack (Renderer process).
- This reduces PostTasks, IPC messages and the number of different
  frame drop callbacks that exist spread out in the code.
- A follow-up CL[1] will delete the frame drop plumbing. This means
  that the only frame drop events will flow in one direction
  (Capturer -> Browser -> Renderer) with no signals back again.

In short, the code in VideoCaptureController is cut-and-paste to
MediaStreamVideoTrack::FrameDeliverer on the video task runner.

"Media.VideoCapture" is renamed to "Media.VideoCapture.Track" to
reflect that collection now happens per-track instead of per-source.

The FrameDeliverer's OnFrameDroppedOnVideoTaskRunner() now handles all
frame drops, regardless if they happened in the capturer or in the
renderer. Instead of notifying the source which resulted in IPC to the
Browser process' VideoCaptureController, the logging and UMA happens in
the FrameDeliverer's LogFrameDroppedOnVideoTaskRunner(). The only IPC
involves is EmitLogMessage() which gets suppressed after 3 logs.

VideoTrackAdapter (the layer between the source and the track which
decides frame rate etc) can also be simplified to having a single
OnFrameDropped that is just forwarded to its callbacks, instead of
mixing callbacks to track and callback to source separately.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4891622

Bug: chromium:1481448
Change-Id: Id375f1ec97da7a2577160a764739eef02f4bf641
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4881385
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Tony Herre <toprice@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1201544}
NOKEYCHECK=True
GitOrigin-RevId: 6c49cd41226991fa73f834e033ce208c5015d322