Don't squash shadows to integer color
DrawLooperBuilder::AddShadow was using an SkColor, thus we were losing
out-of-gamut rgb color for shadows.

crbug.com/1439859 also contains a weird corner case with gradient text
on mac only that requires non-integer font and explicit line height. I
haven't been able to track the cause of it down, but it is worth
separating into a different issue. Here is a test for that:
https://chromium-review.googlesource.com/c/chromium/src/+/4892666

Bug: 1439859
Change-Id: Ifc77c9eff08fd6072e59cdebf3182455e0c0543a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4879833
Reviewed-by: Yi Xu <yiyix@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1201718}
NOKEYCHECK=True
GitOrigin-RevId: 21e1f9f326ac9759e63193f0a193e2b336c3ba3e