view-transition: Add readytorender event
This CL implements the readytorender event for cross-document view
transitions. This event fires within the first 'update the rendering'
steps when a document is presented to the user (regular load, BFCache
activation, prerendering activation).

NotifyViewTransitionRenderingHasBegun is renamed to ActivateFromSnapshot
to match the spec and its call moved to happen within PageAnimator,
after dispatching of this event. Without this change, ViewTransition
moves past the AnimateTagDiscovery state before the first BeginMainFrame
where the readytorender would be dispatched.

This event will also add a hook to get the ViewTransition object but
this will be implemented in a followup CL.

Bug: 1466250
Change-Id: I556fa0210704caba7bdb0eb585dd878235822ad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4832371
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1201822}
NOKEYCHECK=True
GitOrigin-RevId: bc6a41b3452a98c09de942147d6d0d26e15045df