Use common ancestor clip when layer escapes clip of render surface
A layer contributing to a render surface may have a clip state that
is not a descendant of the clip state of the render surface's effect.
This happens at least when a -web-box-reflect element has a
descendant that is not contained by the former. Previously we didn't
apply clip to the render surface, causing the reflection to overflow
ancestor clips, sometimes including the clip of an iframe.

We could change blink to avoid the situation (which would change web
API), but I think we should harden the logic when such situation
occurs. Now behind a kill switch RenderSurfaceCommonAncestorClip,
instead of ignoring all clips, we let the render surface clip to the
common ancestor clip of the clip of the render surface's effect and
the clips of all contributing layers. This can at least prevent
filtered contents from overflowing frames.

Bug: 1478908
Change-Id: If764264447e8a812a86b95d976e6db34467f0b5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4877434
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1201872}
NOKEYCHECK=True
GitOrigin-RevId: 17bfd50ae7118d4c64d26b2508ebe7863a5e9ecb