[css-nesting-ident] Rules that look like custom properties are invalid
After the changes to the css-syntax spec to accommodate relaxed
nesting, we now need to treat any style rule that could be confused
with a custom property declaration as an invalid rule [1].

This CL implements this, however with the tweak we expect to land
in Issue 9336 [2]. The difference vs. the currently specified behavior
is that for top-level ambiguous rules, we only consume the subsequent
block in the error handling. (Otherwise we'd consume the rest of the
stylesheet in these cases, which is a bit severe).

[1] https://drafts.csswg.org/css-syntax/#consume-qualified-rule
[2] https://github.com/w3c/csswg-drafts/issues/9336

Change-Id: If9244c0d2224aa317a3ee0cc95c6b381fdcf78cc
Bug: 1427259
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4894127
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1202382}
NOKEYCHECK=True
GitOrigin-RevId: b38cc34f8084f71d947e416719667aa3d8285e0c