LoAF: ensure scripts are added after microtask checkpoint
In most cases this already works, but there are some edge cases
where the microtask queue can be cleared after the script probe
is executed, for example in TrustedTypes.

To rectify, when adding a script, we first wait unti microtasks are
flushed if they aren't already.

Added a test that reproduces it via trusted-types.

Bug: 1485516
Change-Id: I5bb017ffc441471de65aef7c34ad1dcef159157c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4887255
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Noam Rosenthal <nrosenthal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1202659}
NOKEYCHECK=True
GitOrigin-RevId: c029f30f6d1a5775beb7a708dc48d4b93da0117d