Centralize canvas recording initialization
Before this CL, the recording initialization and the canvas clip stack
restoration was performed "manually", by each and every functions
triggering a flush or a reset of the canvas. This manual maintenance
was error-prone because it was distributing the recording state
maintenance across the code base. New callers of
`recorder_->finishRecordingAsPicture()` for instance would have to
know to restore the recording and do it correctly.

With this change, the `MemoryManagedPaintRecorder`, which owns the
recording, takes care of keeping that recording initialized and valid.

Bug: 1484741
Change-Id: I8243733b4a2866383a907949861990d3ce65d6dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4887595
Reviewed-by: Justin Novosad <junov@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1202734}
NOKEYCHECK=True
GitOrigin-RevId: 2c8290816d919fdd7eecf51f5127da844d2b4a4a