Fix the traversal for dir=auto handling on <slot> to traverse what was intended.
The use of Traversal and FlatTreeTraversal for handling <slot dir=auto>,
prior to this change, starts from the first flat-tree slotted child of
the slot, and from there traverses potentially the entire rest of the
document, ignoring the stay_within argument to the function and the
slotting of children.  This change fixes the traversal to traverse the
children slotted into the <slot> and their light dom descendants.

This is one of two changes needed to fix the failure of:
external/wpt/html/dom/elements/global-attributes/dir-shadow-41.html
in the still-unlanded WPT PR at
https://github.com/web-platform-tests/wpt/pull/29820

Bug: 576815
Change-Id: Ifc0277cff4d94a30e2dd932004a55ed54bfa60bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4805164
Reviewed-by: Di Zhang <dizhangg@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1202823}
NOKEYCHECK=True
GitOrigin-RevId: 0e041703a83020d5e945c963b87af53ae75297a6