Add automatic beacon histograms.
To better assist in debugging a discrepancy we're noticing between
reports from existing click reporting APIs and the
`reserved.top_navigation` automatic beacon API, we are adding metrics
to get a better sense of why some clicks are not resulting in automatic
beacons being sent out.

This CL adds a histogram to log the outcome of an attempt to send an
automatic beacon. It tracks if the beacon is successfully passed off to
the `SendFencedFrameReportingBeaconInternal()` function, or if it was
not passed due to one of the following reasons:
- There was no transient user activation when the top-level navigation
  happened.
- There was no automatic beacon information in place.
- The frame initiating the top-level navigation is not same-origin to
  the mapped url in the fenced frame properties.
- No destinations were registered in the automatic beacon info.

This CL adds histograms to log if a top-level navigation originating
from a fenced frame is started, as well as if it commits. This is being
added to see if there is a percentage of top-level navigations that
originate from a fenced frame that result in some sort of failure (such
as a 404, a timeout, or the load being blocked by a content blocker).

Change-Id: Ie24ee3a2b0e3d98e10fbd5844b85f024fd363d32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4877250
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Cr-Commit-Position: refs/heads/main@{#1202851}
NOKEYCHECK=True
GitOrigin-RevId: 42f47fc04f40bd840b8c33fb9ac61226b7b79137