Let PLSA::UsesCompositedScrolling() use the current compositing status
This makes it always use the CompositeScrollAfterPaint way for the
function. This prepares for removing pre-CompositeScrollAfterPaint
code. It's safe to do it now because the function is used in only
two places where the risk of breakage is low:
1. ScrollableArea/PLSA, to show non-composited overlay scrollbars.
2. Animation, to check if we can start a scroll-driven animation
   in the compositor, behind a flag ScrollTimelineOnCompositor.

For 1, now we will show overlay scrollbar in the main thread iff
the scrollbar is currently composited, which is more accurate.
The change also fixes potential bug that in CompositeScrollAfterPaint
we may show scrollbars unnecessarily because of the previous check
for NeedsCompositedScrolling which only returns true if the scroller
is forced to be composited in CompositeScrollAfterPaint.

For 2, the new code is more accurate, if there is any change.
We may also remove the check soon because we are already allowing
composited scroll-driven animation in cc (https://crrev.com/1136319).

Bug: 1414885
Change-Id: Ied719dd5438651d65eb8bfecd471455fcc06d6a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4898509
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1202994}
NOKEYCHECK=True
GitOrigin-RevId: 505a2ab0e1464257ba0389d718e07e6987fa554d