Update the aspect ratio of the created font for font-size-adjust: from-font
When resolving styles, we use the parent-style font to compute the
aspect ratio for font-size-adjust: from-font because the font style of
the current element could be unresolved or unloaded. Unfortunately, this
causes the failure of font-size-adjust-014.html. When we load a new font
from the font-family, we need to update the aspect value with the used
font's actual font metrics to make the from-font value work.

Test: external/wpt/css/css-fonts/font-size-adjust-014.html

Bug: 1219875, 626703
Change-Id: I5298ecdd12687c9f015f25eeb3c02ed8dec72d8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4867139
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: ChangSeok Oh <changseok.oh@bytedance.com>
Cr-Commit-Position: refs/heads/main@{#1203130}
NOKEYCHECK=True
GitOrigin-RevId: b18671a0bbb6a2d4d4e39fc32bd1b1ab175c03a5